Bugs Fixed, Optimization and etc..
