+++ PetShelter/PetShelter/AnimalDatabase.bin
@@ -30,14 +30,6 @@ Aggresive
 2
 21
 75
-Dog
-Chihuahua
-Aggresive
-0
-1
-15
-26
-85
 Cat
 Toyger
 Lazy
@@ -70,6 +62,14 @@ Lazy
 8
 30
 90
+Cat
+Toyger
+Grumpy
+0
+0
+5
+23
+100
 Rabbit
 Cottontail
 Lazy
@@ -150,11 +150,3 @@ Peacefull
 5
 29
 30
-Cat
-Toyger
-Grumpy
-0
-0
-5
-23
-100
